First auto deploy attempt from the webhook [no deploy]

+++ devel/sync_lock.sh
@@ -6,7 +6,10 @@ then
 fi
 wait_for_command.sh devstats 3600 || exit 2
 cronctl.sh devstats off || exit 3
-wait_for_command.sh webhook 600 || exit 4
-cronctl.sh webhook off || exit 5
-killall webhook
+if [ -z "$FROM_WEBHOOK" ]
+then
+  wait_for_command.sh webhook 600 || exit 4
+  cronctl.sh webhook off || exit 5
+  killall webhook
+fi
 echo 'All sync and deploy jobs stopped and disabled'
